pyproject.toml
[tool.poetry]
name = "dbgear"
version = "0.1.0"
description = ""
authors = ["tamuto <tamuto@infodb.jp>"]
readme = "README.md"
homepage = "https://github.com/tamuto/dbgear"
repository = "https://github.com/tamuto/dbgear"

packages =[
    { include="dbgear", from="." }
]

[tool.poetry.scripts]
dbgear = "dbgear.main:execute"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = "^0.14.0"
aiofiles = "^0.7.0"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
